Added token system to admin Bookings and Documents

+++ src/main/java/com/example/demo/controller/BookingController.java
@@ -4,9 +4,7 @@ package com.example.demo.controller;
 import com.example.demo.entity.Booking;
 import com.example.demo.entity.document.Document;
 import com.example.demo.entity.user.User;
-import com.example.demo.exception.DocumentNotFoundException;
-import com.example.demo.exception.NullIdException;
-import com.example.demo.exception.UserNotFoundException;
+import com.example.demo.exception.*;
 import com.example.demo.repository.BookingRepository;
 import com.example.demo.service.BookingService;
 import com.example.demo.service.DocumentService;
@@ -32,7 +30,10 @@ public class BookingController {
     }
 
     @GetMapping("/booking/find")
-    public Iterable<Booking> findBookingByUserId(@RequestParam(name = "id", defaultValue = "-1") Integer id) {
+    public Iterable<Booking> findBookingByUserId(@RequestParam(name = "id", defaultValue = "-1") Integer id, HttpServletRequest request) {
+        ParserToken token = TokenAuthenticationService.getAuthentication(request);
+        if (token == null) throw new UnauthorizedException();
+        if (!token.role.equals("admin")) throw new AccessDeniedException();
         return bookingService.findAll()
                 .stream()
                 .filter(booking -> booking.getUser().getId().equals(id))
@@ -40,8 +41,24 @@ public class BookingController {
                 .collect(Collectors.toList());
     }
 
+    @GetMapping("/booking/findself")
+    public Iterable<Booking> findMyBooking(HttpServletRequest request) {
+        ParserToken token = TokenAuthenticationService.getAuthentication(request);
+        if (token == null) throw new UnauthorizedException();
+        if (token.role.equals("admin")) throw new AccessDeniedException();
+        return bookingService.findAll()
+                .stream()
+                .filter(booking -> booking.getUser().getId().equals(token.id))
+                .filter(booking -> !booking.isHasBackRequest())
+                .collect(Collectors.toList());
+    }
+
     @GetMapping("/booking/findback")
-    public Iterable<Booking> findReturnBooks() {
+    public Iterable<Booking> findReturnBooks(HttpServletRequest request) {
+        ParserToken token = TokenAuthenticationService.getAuthentication(request);
+        if (token == null) throw new UnauthorizedException();
+        if (!token.role.equals("admin")) throw new AccessDeniedException();
+
         return bookingService.findAll()
                 .stream()
                 .filter(booking -> !booking.isClose())
@@ -50,18 +67,22 @@ public class BookingController {
     }
 
     @GetMapping("/booking/findall")
-    public Iterable<Booking> findAllBookings() {
+    public Iterable<Booking> findAllBookings(HttpServletRequest request) {
+        ParserToken token = TokenAuthenticationService.getAuthentication(request);
+        if (token == null) throw new UnauthorizedException();
+        if (!token.role.equals("admin")) throw new AccessDeniedException();
         return bookingService.findAll();
     }
 
     @PostMapping("/booking/request")
-    public void requestDocumentById(@RequestParam(value = "id", defaultValue = "") Integer id, HttpServletRequest request) {
+    public void requestDocumentById(@RequestParam(value = "id", defaultValue = "") Integer documentId, HttpServletRequest request) {
         ParserToken token = TokenAuthenticationService.getAuthentication(request);
         if (token == null)
-            throw new NullIdException();//TODO: TOKEN EXCEPTION
-        if (id == null)
+            throw new UnauthorizedException();
+        if (token.role.equals("admin")) throw new AccessDeniedException();
+        if (documentId == null)
             throw new NullIdException();
-        Document document = documentService.findById(id);
+        Document document = documentService.findById(documentId);
         if (document == null)
             throw new DocumentNotFoundException();
         if (token.id == null)
@@ -71,9 +92,11 @@ public class BookingController {
             throw new UserNotFoundException();
         if (!document.isReference() && document.getCount() > 0) {
             bookingService.save(new Booking(user, document, null, false, 0, false));
+            document.setCount(document.getCount() - 1);
+            documentService.save(document);
         }
         else{
-            throw new DocumentNotFoundException();//TODO: REFERENCE/NO BOOK EXCEPTION
+            throw new AccessDeniedException();
         }
     }
 
@@ -81,35 +104,52 @@ public class BookingController {
     public void returnDocumentById(@RequestParam Integer id, HttpServletRequest request){
         ParserToken token = TokenAuthenticationService.getAuthentication(request);
         if (token == null)
-            throw new NullIdException();//TODO: TOKEN EXCEPTION
+            throw new UnauthorizedException();
+        if (token.role.equals("admin")) throw new AccessDeniedException();
         if (id == null)
             throw new NullIdException();
         Booking booking = bookingService.getBookingById(id);
         if (booking == null)
-            throw new NullIdException();//TODO: BOOKING EXCEPTION
+            throw new BookingNotFoundException();
         booking.setHasBackRequest(true);
         bookingService.save(booking);
     }
 
     @PutMapping("/booking/close")
-    public void closeBooking(@RequestParam Integer id){
+    public void closeBooking(@RequestParam Integer id, HttpServletRequest request){
+        ParserToken token = TokenAuthenticationService.getAuthentication(request);
+        if (token == null)
+            throw new UnauthorizedException();
+        if (!token.role.equals("admin")) throw new AccessDeniedException();
+
         Booking booking = bookingService.getBookingById(id);
         if(booking == null){
-            throw new NullIdException();//TODO: BOOKING EXCEPTION
+            throw new BookingNotFoundException();
         }
         booking.setClose(true);
         bookingService.save(booking);
+        Document document = booking.getDocument();
+        document.setCount(document.getCount() + 1);
+        documentService.save(document);
     }
 
     @Transactional
     @DeleteMapping("/booking/remove")
-    public void removeBooking(@RequestBody Booking booking){
+    public void removeBooking(@RequestBody Booking booking, HttpServletRequest request){
+        ParserToken token = TokenAuthenticationService.getAuthentication(request);
+        if (token == null)
+            throw new UnauthorizedException();
+        if (!token.role.equals("admin")) throw new AccessDeniedException();
         this.bookingService.removeBookingById(booking.getId());
     }
 
     @Transactional
     @DeleteMapping("/booking/idremove")
-    public void removeBookingById(@RequestParam Integer id){
+    public void removeBookingById(@RequestParam Integer id, HttpServletRequest request){
+        ParserToken token = TokenAuthenticationService.getAuthentication(request);
+        if (token == null)
+            throw new UnauthorizedException();
+        if (!token.role.equals("admin")) throw new AccessDeniedException();
         this.bookingService.removeBookingById(id);
     }
 }
